fix(assignment): satisfy protected source governance
