+++ framework/core/tests/python/test_coordination_locks.py
@@ -23,7 +23,7 @@ _LOCKS_PATH = (
     / "python"
     / "kungfu"
     / "coordination"
-    / "__init__.py"
+    / "locks.py"
 )
